Add files via upload
PCB for the MOSFETs controlling LED strips and the FANs. The MOSFETs I use are IRL8721, these can handle 16A so they wont break that easy. The have the TO-220 footprint so if you have or want to use other MOSFETs that also have a TO-220 footprint you can just put them in.