Add files via upload
This is a separate power supply board to power the LEDs/FANs and the Arduino. For stability and safety I would make 2 of these, one with a 9v dc hi-link module to power the Arduino and one with the HLK 20M12, the 12v hi-link module for the LEDs/FANs. You could also just use one 12v for all, but thats at your own risk. Together with the updated code this should solve most of the stability issues also!